move read_mapper and input_data_storage into utils
