+++ isoquant_lib/utils/input_data_storage.py
