+++ isoquant_lib/utils/read_mapper.py
@@ -15,9 +15,9 @@ import pysam
 import gzip
 import sys
 
-from .common import get_path_to_program
+from isoquant_lib.common import get_path_to_program
 from isoquant_lib.utils.error_codes import IsoQuantExitCode
-from .gtf2db import convert_db_to_gtf, db2bed
+from isoquant_lib.gtf2db import convert_db_to_gtf, db2bed
 from .input_data_storage import SampleData, InputDataType
 
 logger = logging.getLogger('IsoQuant')
